#86d10mu1p [IN-PROGRESS]: MAP, add table with bars and heat map
