+++ js/packages/react-ui/src/components/Table/TableWithBars/index.ts
