+++ js/packages/react-ui/src/components/Table/TableWithHeatmap/index.ts
